pyproject.toml
[tool.poetry]
name = "pastabean-auth-service"
version = "0.1.0"
description = ""
authors = ["Daniil-Solo <daniil.solo1723@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
sqlalchemy = "^2.0.28"
uvicorn = "^0.28.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.2"
pytest = "^8.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"